Only calculate difficulties on import for now.

+++ osu.Game/Beatmaps/Drawables/BeatmapPanel.cs
@@ -138,7 +138,7 @@ namespace osu.Game.Beatmaps.Drawables
                                 },
                                 starCounter = new StarCounter
                                 {
-                                    Count = beatmap.StarDifficulty,
+                                    Count = (float)beatmap.StarDifficulty,
                                     Scale = new Vector2(0.8f),
                                 }
                             }
